refactor: factor out popover + columns
